+++ src/utils/builtMuiTheme.ts
@@ -1,16 +1,15 @@
 import {MuiTheme} from 'react-openfin';
-import { createMuiTheme } from '@material-ui/core/styles';
+import createMuiTheme, { ThemeOptions } from '@material-ui/core/styles/createMuiTheme';
 import initState from '../assets/jss/openfin-starter-constant';
 
-export const buildMuiTheme = (theme:MuiTheme)=>{
+export const buildMuiTheme = (options:ThemeOptions, theme?:MuiTheme)=>{
     return createMuiTheme({
         typography:{
-            useNextVariants:true,
             fontSize: 8,
             htmlFontSize: 10,
         },
         palette:{
-            type:theme,
+            type:theme?theme:void 0,
             primary:{
                 main:initState.primaryColor,
             },
@@ -23,20 +22,10 @@ export const buildMuiTheme = (theme:MuiTheme)=>{
         },
         overrides:{
             MuiButton:{
-                root:{
-                    color:'white',
-                },
                 contained:{
                     padding:'0px 5px',
                     minWidth:'50px',
                     minHeight:'20px',
-                    color:'white',
-                },
-                containedPrimary:{
-                    color:'white',
-                },
-                containedSecondary:{
-                    color:'white',
                 },
                 sizeSmall:{
                     padding:'0px 5px',
@@ -91,12 +80,9 @@ export const buildMuiTheme = (theme:MuiTheme)=>{
                 root:{
                     minHeight:'24px',
                 },
-                labelContainer:{
-                    paddingTop:'0px',
-                    paddingBottom:'0px',
-                }
             },
-        }
+        },
+        ...options,
     })
 };